fix: globals.css에서 swiper 스타일 제거 및 카드/리뷰카드 타입 분리, 깃 리뷰 반영 코드 수정(SCENTLY-77)
- globals.css에서 Swiper 관련 기본 스타일 삭제로 스타일 충돌 방지
- Card 컴포넌트와 ReviewCard 컴포넌트의 타입을 별도 파일로 분리하여 관리 및 재사용성 향상
- 깃 코드 리뷰 피드백을 반영하여 컴포넌트 prop 타입, 불필요한 props 제거, 에러 핸들링 및 코드 정리
- 전반적인 코드 품질 향상 및 유지보수 용이성 개선
