+++ src/app/globals.css
@@ -1,5 +1,4 @@
 @import "tailwindcss";
-@import "swiper/swiper-bundle.css";
 
 @theme {
   /* 높이 값 */
