+++ src/components/common/card-swiper/CardSwiper.types.ts
@@ -1,10 +1,11 @@
-import { CardProps } from "@components/common/Card";
+// src/components/common/card-swiper/CardSwiper.types.ts
+import type { CardProps } from "@components/common/card/Card.type";
 
 export type CardSwiperProps = {
   withNavigation?: boolean;
   withPagination?: boolean;
-  slidesPerView?: number;
   spaceBetween?: number;
-  items: CardProps[]; // 카드 데이터 배열
+  items: CardProps[];
+  autoplay?: boolean;
   className?: string;
 };
