+++ src/components/common/card/Card.tsx
@@ -1,3 +1,4 @@
+// src/components/common/card/Card.tsx
 "use client";
 
 import Checkbox from "@components/ui/input/Checkbox";
@@ -6,35 +7,9 @@ import { useState } from "react";
 import Image from "next/image";
 import { cn } from "@utils/cn";
 
-export type CardProps = {
-  handleHeartChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
-  imageUrl?: string;
-  isLiked: boolean;
-  tags?: string[];
-  price: number;
-  name: string;
-};
+import type { CardProps } from "./Card.type";
 
 const FALLBACK_IMAGE = "/fallback-image.svg";
-
-/** Card 컴포넌트
- * @param tags - 문자열 배열. []로 설정한 이유는 부모 컴포넌트에서 tags를 전달하지 않아도 오류 없이 작동하도록하기 위함.
- * @param imageUrl - 카드에 표시할 이미지 경로
- * @param price - 숫자 타입의 가격
- * @param name - 상품 이름 또는 카드 제목 역할
- * @param isLiked - 사용자가 이 상품에 '좋아요'를 눌렀는지 여부
- * @param handleHeartChange - 좋아요(하트) 상태가 변경될 때 호출되는 콜백 함수
- *
- * @example
- * <Card
- *   tags={['타이틀 이름름','타이틀 이름2']}
- *   imageUrl="/images/product.png"
- *   price={19990}
- *   name="상품 이름"
- *   isLiked={true}
- *   handleHeartChange={(e) => {}}
- * />
- */
 
 const Card = ({
   handleHeartChange,
@@ -52,7 +27,7 @@ const Card = ({
       <div className="relative mb-2 flex h-[250px] w-full items-center justify-center overflow-hidden rounded-lg border border-border-default">
         <Checkbox
           className="absolute top-[16px] right-[16px] m-2 h-[32px] w-[32px]"
-          onChange={handleHeartChange}
+          onChange={(e) => handleHeartChange && handleHeartChange(e)}
           checked={isLiked}
           type="heart"
           name="heart"
@@ -60,7 +35,7 @@ const Card = ({
         />
         <Image
           className={cn(
-            imageUrl ? "obeject-contain h-full w-full" : "object-cover",
+            imageUrl ? "h-full w-full object-contain" : "object-cover",
           )}
           onError={() => setImgSrc(FALLBACK_IMAGE)}
           src={imgSrc ?? FALLBACK_IMAGE}
