+++ src/components/common/review-card/ReviewCard.types.ts
@@ -1,0 +1,12 @@
+export type ReviewCardProps = {
+  productImage?: string;
+  productPrice: string;
+  productName: string;
+  imageUrl?: string;
+  timeAgo: string;
+  rating: number;
+  review: string;
+  writer: string;
+  date: string;
+  id: string;
+};
