+++ src/components/common/review-swiper/ReviewSwiper.types.ts
@@ -1,10 +1,9 @@
-import type { ReviewCardProps } from "@components/common/ReviewCard";
+import type { ReviewCardProps } from "@components/common/review-card/ReviewCard.types";
 
 export type ReviewSwiperProps = {
   items: ReviewCardProps[];
   withNavigation?: boolean;
   withPagination?: boolean;
-  slidesPerView?: number;
   spaceBetween?: number;
   className?: string;
   autoplay?: boolean;
